Refactor
This removes the worker pool entirely. The application is not cpu bound,
we spend most of our time waiting for an http request or external
program (ffmpeg, livestreamer, etc) so the worker pool doesn't gain us
anything.

This uses promises to chain together the four async tasks formerly
in worker.js. This removes the need for timeouts and should make the application
considerably more robust.