Adjust crop area for test server and 1.0
The box containing the number of players alive has shifted slightly
with the test server (and 1.0 pubg release on the 12/20/2017). This
will correct the cropScreenshots function to center over this new
location.

We might not want to merge yet, but most people streaming PUBG will
be streaming the test server until the 1.0 release, and this will
need to be fixed anyway.

Signed-off-by: Cullen Taylor <cullentaylor@outlook.com>