Update the OCR to use Watson Visual Recognition service
Uses Watson visual recognition service to ocr the screen. It finds
the number of players left alive by finding the integer between
(0,100] that is closest the word alive. This is helpful because
it is less vulnerable to game updates and automatically will
eliminate streamers not playing the English version of the game.

Signed-off-by: pwplusni@us.ibm.com
Signed-off-by: rdayao@us.ibm.com