Update package.json to reflect new npm package
I went ahead and published rotisserie to npm. However, the name
rotisserie was taken on npmjs, is deprecated, but seems to still be
kept around for unknown reasons. This will update the name of the
package in package.json as well as the npm badge in the readme, as it
previously pointed to the old pubgredzone package which is super old.

Signed-off-by: Cullen Taylor <cullentaylor@outlook.com>