Add ibm metrics collector client
This will make use of the IBM Metrics Collector Client, which seems
to be a desired thing for all of IBM's code patterns to track our
kubernetes deployments. I will presumably need to shutdown/migrate
our google analytics stuff anyway.

The information on this should stay static, so it shouldn't need to
be templated any further.

Reference:
https://github.com/IBM/metrics-collector-client-kubernetes

Signed-off-by: Cullen Taylor <mctaylor@us.ibm.com>