Refactor stream processing
Rotisserie now depends on having an instance of Redis running
in the cluster. Redis is used to store the list of streams
gathered by the application. Those streams are processed by
a set of workers who capture clips, take screenshots and pass
those screenshots to the OCR service. The OCR results are added
to a different Redis key. The main application checks that results
key in Redis and uses that to find the lowest stream.

Because the workers are decoupled from the collection of the stream
information they can be scaled separately.