Fix vulnerable dependencies
GitHub is yelling about a dependency vulnerability again. This will
upgrade mime to 1.4.1 to get rid of that and also apply the magic
of npm's auditing to upgrade packages it also found to have
vulnerabilities.

Really old (2.x.x) version(s) of hoek are also still being pulled in
by other packages, which are the only moderate vulnerabilities left.